% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bs.R
\name{get_bs}
\alias{get_bs}
\title{get_bs}
\usage{
get_bs(ticker, report_type = "annual")
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}

\item{report_type}{Accepts "quarterly" or "annual". Default is "annual". Both return 4 intervals.}
}
\value{
Dataframe
}
\description{
A tidy dataframe of balance sheet data.
}
\examples{
get_bs("AAPL")

get_bs(c("AAPL", "MSFT"), report_type = "quarterly")
}
