% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_company_names}
\alias{get_company_names}
\title{get_company_names}
\usage{
get_company_names(ticker)
}
\arguments{
\item{ticker}{The vector of tickers to get the Company names of.}
}
\value{

}
\description{
get_company_names
}
\examples{
get_company_names(c("MSFT", "AAPL"))
}
