% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_financials.R
\name{get_financials}
\alias{get_financials}
\title{get_financials}
\usage{
get_financials(ticker, report_type = "annual")
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}

\item{report_type}{Accepts "quarterly" or "annual". Default is "annual". Both return 4 intervals.}
}
\value{
Dataframe
}
\description{
A single tidy dataframe of income statement, balance sheet, and cash flow data.
}
\examples{
get_financials("AAPL")

get_financials(c("AAPL", "MSFT"), report_type = "quarterly")

}
