% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_income.R
\name{get_income}
\alias{get_income}
\title{get_income}
\usage{
get_income(ticker, report_type = "annual")
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}

\item{report_type}{Accepts "quarterly" or "annual". Default is "annual". Both return 4 intervals.}
}
\value{
Dataframe
}
\description{
A tidy dataframe of income statement data.
}
\examples{
get_income("AAPL")

get_income(c("AAPL", "MSFT"), report_type = "quarterly")
}
