% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price.R
\name{get_price}
\alias{get_price}
\title{get_price}
\usage{
get_price(ticker)
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}
}
\value{
Dataframe
}
\description{
Returns price, market cap and other similar data.
}
\examples{

get_price("MSFT")
get_price(c("AAPL", "GOOG"))
}
