% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summaries.R
\name{get_summaries}
\alias{get_summaries}
\title{get_summaries}
\usage{
get_summaries(ticker)
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}
}
\value{
Dataframe
}
\description{
get_summaries
}
