% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBS.R
\name{getBS}
\alias{getBS}
\title{getBS}
\usage{
getBS(ticker, report_type = "annual")
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}

\item{report_type}{Accepts "quarterly" or "annual". Default is "annual". Both return 4 intervals.}
}
\value{
Dataframe
}
\description{
A tidy dataframe of balance sheet data.
}
\examples{
getBS("AAPL")

getBS(c("AAPL", "MSFT"), report_type = "quarterly")
}
