% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCF.R
\name{getCF}
\alias{getCF}
\title{getCF}
\usage{
getCF(ticker, report_type = "annual")
}
\arguments{
\item{ticker}{A string vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}

\item{report_type}{Accepts "quarterly" or "annual". Default is "annual". Both return 4 intervals.}
}
\value{
Dataframe
}
\description{
A tidy dataframe of cash flow data.
}
\examples{
getCF("AAPL")

getCF(c("AAPL", "MSFT"), report_type = "quarterly")
}
