% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinancials.R
\name{getFinancials}
\alias{getFinancials}
\title{getFinancials}
\usage{
getFinancials(ticker, report_type = "annual")
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}

\item{report_type}{Accepts "quarterly" or "annual". Default is "annual". Both return 4 intervals.}
}
\value{
Dataframe
}
\description{
A single tidy dataframe of income statement, balance sheet, and cash flow data.
}
\examples{
getFinancials("AAPL")

getFinancials(c("AAPL", "MSFT"), report_type = "quarterly")

}
