% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIncome.R
\name{getIncome}
\alias{getIncome}
\title{Get Income Statement}
\usage{
getIncome(ticker, format = "tidy", assign = TRUE)
}
\arguments{
\item{ticker}{The stock ticker of the company}

\item{format}{Choose between "raw", "clean" and "tidy" format. For readability choose "raw" and for usability in R, "tidy", which is also the default.}

\item{assign}{If TRUE, it will assign the function output dataframe to "ticker_income". If FALSE, the function will only return the dataframe, and it's up to the user to assign it to the environment. Default is TRUE.}
}
\value{
A dataframe with income statement data from the last 4 yearly reports.
}
\description{
Get Income Statement
}
\examples{
getIncome("AAPL")
getIncome("KO", format="raw")
getIncome("MSFT", format="clean", assign=FALSE)
}
