% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrice.R
\name{getPrice}
\alias{getPrice}
\title{getPrice}
\usage{
getPrice(ticker)
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}
}
\value{
Dataframe
}
\description{
Returns price, market cap and other similar data.
}
\examples{

getPrice("MSFT")
getPrice(c("AAPL", "GOOG"))
}
