% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaries.R
\name{getSummaries}
\alias{getSummaries}
\title{getSummaries}
\usage{
getSummaries(ticker)
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}
}
\value{
Dataframe
}
\description{
getSummaries
}
