% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar.R
\name{progress_bar}
\alias{progress_bar}
\title{progress_bar}
\usage{
progress_bar(datatype, ticker)
}
\arguments{
\item{datatype}{The type of data that is accessed.}

\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}
}
\value{
Progress Bar.
}
\description{
progress_bar
}
