% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_download.R
\name{safe_download}
\alias{safe_download}
\title{safe_download}
\usage{
safe_download(ticker, report_type, proto_function)
}
\arguments{
\item{ticker}{A character vector of stock tickers. Ex. ticker = c("AAPL", "MSFT")}

\item{report_type}{Either "quarterly" or "annual", depending whether the proto function accepts such an argument.}

\item{proto_function}{A function that needs to run safely and not stop at errors.}
}
\value{
Dataframe with errors as attributes.
}
\description{
safe_download
}
