% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_stock.R
\name{search_stock}
\alias{search_stock}
\title{search_stock}
\usage{
search_stock(search_term)
}
\arguments{
\item{search_term}{The name, cusip code or stock ticker of a security.}
}
\value{
Search results of securities matching the query. Available data: exchange, shortname, ticker symbol, security type and more.
}
\description{
search_stock
}
