% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_names.R
\name{get_company_names}
\alias{get_company_names}
\title{get_company_names}
\usage{
get_company_names(ticker)
}
\arguments{
\item{ticker}{A vector of stock tickers. Ex. c("MSFT", "AAPL")}
}
\value{

}
\description{
Given a vector of tickers, returns a vector of company names.
}
\examples{
get_company_names(c("MSFT", "AAPL"))
}
