% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinancials.R
\name{getFinancials}
\alias{getFinancials}
\title{Get Financial Statements}
\usage{
getFinancials(ticker, assign = TRUE)
}
\arguments{
\item{ticker}{The stock ticker of the company as a string, or a vector of stock tickers as strings.}

\item{assign}{Choose between TRUE or FALSE. If TRUE, function output dataframe is assigned as "financials" to the environment. If FALSE, the function will only return the dataframe and it's up to the user to save it under a name. FALSE is advised when getting financial data for several sets of companies, as data will be overwritten when TRUE. Default is TRUE.}
}
\value{
A dataframe with all financial data from the last 4 yearly reports.
}
\description{
Get Financial Statements
}
\examples{
getFinancials("AAPL")
getFinancials(c("AAPL", "MSFT"), assign=FALSE)
}
