% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinancials.R
\name{getFinancials}
\alias{getFinancials}
\title{Get Financial Statements}
\usage{
getFinancials(ticker, assign = TRUE)
}
\arguments{
\item{ticker}{The stock ticker of the company as a string, or a vector of stock tickers as strings.}

\item{assign}{Choose between TRUE or FALSE. If TRUE, the output dataframe is assigned as "financials" to the environment. If FALSE, the function will only return the dataframe and it's up to the user to save it under a name. FALSE is advised when running getFinancials several times, as the previous output dataframes will be overwritten when assign=TRUE. Default is TRUE.}
}
\value{
A dataframe with all financial data from the last 4 yearly reports.
}
\description{
Returns a single tidy dataframe with historical data from the income statement, balance sheet, and cash flow statement.
}
\examples{
getFinancials("AAPL")
getFinancials(c("AAPL", "MSFT"), assign=FALSE)
}
