% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_download.R
\name{safe_download}
\alias{safe_download}
\title{safe_download}
\usage{
safe_download(vector, proto_function, ...)
}
\arguments{
\item{vector}{A vector of things to be iterated over. Ex. ticker = c("AAPL", "MSFT")}

\item{proto_function}{A function that needs to run safely and not stop at errors.}

\item{...}{Any other parameters that the proto_function accepts.}
}
\value{
Dataframe with errors as attributes.
}
\description{
safe_download
}
