% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_stock.R
\name{search_stock}
\alias{search_stock}
\title{search_stock}
\usage{
search_stock(search_term, keep_results = "all")
}
\arguments{
\item{search_term}{The name, cusip code, isin code or stock ticker of a security.}

\item{keep_results}{Set "top" for the first result per search, or "all" for all search results.}
}
\value{
Search results of securities matching the query. Available data: exchange name, shortname, ticker symbol, security type and more.
}
\description{
Search for stocks and securities using company names, isin and cusip codes or tickers. Useful to get the exact tickers yahoo finance uses in order to query other financial data.
}
\examples{
search_stock("Nike", keep_results = "all")
search_stock(c("Nike", "adidas"), keep_results = "top")
search_stock(search_term = "US4581401001", keep_results = "top")
search_stock(search_term = "594918104", keep_results = "top")
}
